deep_equation/requirements.txt
tqdm==4.58.0
pytest
Pillow
torch
torchvision
numpy